pyproject.toml
[project]
name = "mas-for-multimodal-knowledge-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.47.0",
    "langchain-community",
    "langchain-huggingface>=1.0.0",
    "marker-pdf[full]>=1.10.1",
    "markitdown[all]>=0.1.3",
    "matplotlib>=3.10.7",
    "neo4j-graphrag>=1.10.1",
    "pyprojroot>=0.3.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
]
